[refactor] exporting all location list tile to dedicated component
