+++ lib/utils/information/InformationSheet.dart
@@ -1,14 +1,13 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_i18n/flutter_i18n.dart';
-import 'package:pandemia/utils/information/GeoSwitch.dart';
+import 'package:pandemia/utils/information/GeoTile.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 /// Displays a text sheet displaying information about the application.
 class InformationSheet {
   BuildContext _context;
   EdgeInsets _padding = EdgeInsets.symmetric(horizontal: 20, vertical: 10);
-  GeoSwitch _switch = new GeoSwitch();
   InformationSheet(this._context);
 
   void show () {
@@ -40,18 +39,7 @@ class InformationSheet {
                       FlutterI18n.translate(_context, "home_info_data_text1")
                   ),
                 ),
-                new ListTile(
-                  onTap: () => _switch.toggle(),
-                  leading: new Icon(Icons.location_on),
-                  contentPadding: _padding,
-                  title: new Text(
-                      FlutterI18n.translate(_context, "home_info_location_title")
-                  ),
-                  subtitle: Text(
-                      FlutterI18n.translate(_context, "home_info_location_text")
-                  ),
-                  trailing: _switch,
-                ),
+                GeoTile (),
                 new ListTile(
                   leading: new Icon(Icons.warning),
                   contentPadding: _padding,
